Add php 8 syntax support
The new under development version of our reflection lib adds
supports php8 syntax. To be able to support a number of more complex
reflection types like ConstructorPropertyPromotion required some internal
changes to the lib. This required a new way of configuring the strategies.

That's why I introduced the CompilerPass to make it easier to configure new strategies
but also to be more flexible when new strategies are introduced in the lib.