Rework class overview and add search tests
In this change, I had to refactor the class overview tests because the
previous version was broken. To be fair, no idea how it was possible
that test passed?

In addition, I have extracted the search tests to an include file; so
that we can execute that test on every page's test set without having to
copy paste everything