Remove environment expose from DocumentNode
To be able to cache the DocumentNode we need to remove all state that
is not related to the Node from the nodes. Environment is not required
for the document node and should not be exposed to the outside world.

This is a first step to remove the environment form the DocumentNode.
More will follow.