Move documents in do GuideDocumentationSet
To be able to use the compiler to generate the TOC we need to have
the Documents from guides in our descriptors. This commit removes
the global state Documents set from the application. And stores the documents
in the descriptors which makes it possible to render multiple sets of documents.