Remove callables from nodes to make them cachable.
Callables can not be cached. This is an issue for phpDocumentor since
we want to use caching heavy after parsing to speed up the process of
re-running phpdocumentor. Therefor the callbacks are replaced with an new
node type which will be rendered by a new type of renderer.
This might be an in between step to more specialized renderers. But for now
this will work.