Introduce TOC builder in compiler stage
To allow all descriptors to provide content to the table of contents
I introduced the TableOfContentsBuilder which will iterate over the
namespaces, packages and guide documents to add them in a TOC on each
documentationset.

Besides that all descriptors could be included this allows us to share
the TOC between documentationsets and versions when applicable.
The TOC rendering replaces the hard coded menu's in the templates. making
it easier to render them.