PHP 8.1: fix various deprecation warnings
A run on PHP 8.1 currently shows various notices along the lines of:
```
Deprecated: Return type of [CLASSNAME]::[METHODNAME]() should either be compatible with [INTERFACENAME]::[METHODNAME](): [Type], or the #[ReturnTypeWillChange] attribute should be used to temporarily suppress the notice
```

These deprecation notices relate to the [Return types for internal methods RFC](https://wiki.php.net/rfc/internal_method_return_types) in PHP 8.1.

Basically, as of PHP 8.1, these methods in classes which implement PHP native interfaces are expected to have a return type declared.
The return type can be the same as used in PHP itself or a more specific type. This complies with the Liskov principle of covariance, which allows the return type of a child overloaded method to be more specific than that of the parent.

While this package has a minimum PHP requirement of PHP 7.2 based on the `composer.json` file, several of these return types expect for the PHP 8.0 `mixed` type to be used, so for consistency across the codebase, I've used the attribute for now to silence the warning.

While attributes are a PHP 8.0 feature only, due to the syntax choice for `#[]`, they will ignored in PHP < 8 and can be safely added.