Fix issue with link rendering
The environment is used to render links. These links are collected during
the parsing process of each document, and stored in the global enviroment which belongs to the
document.

When we splitted the parsing process from the rendering to make the result cachable, the environment
was destroyed after the parsing process was finished.

Because the rendering is still using this environment we had to recreate the object on render. Which
doesn't have the required links anymore. By adding them for now I restored the original behavior, where
the environment has all links.