Add support for php:class role in Guides
In a guides, we would like to refer to an API element in the same
project. Although references were added for this, Domain support was not
added in the parshing and handling of references. In this change I have
introduced the concept of Domains in reference parsing and changes the
different php roles to work.

See
https://www.sphinx-doc.org/en/master/usage/restructuredtext/domains.html
for more information on Domains.