Restore functionality of the 'replace' directive
In the process of splitting the parsing from the rendering, we did not
take into account that variables were stored on the Environment class.
In the original implementation this Environment class shared state
between just about everything in the whole app.

By pushing the variables into the DocumentDescriptor and restoring it
when rendering, the replace functionality now works again.

Additionally, I found a slight bug in the DocumentParser; when a
Directive is followed by something other than a CodeNode we want to
ignore it and not pass it as a SpanNode to the directive's process
function. In that scenario, the deeper business logic will convert that
SpanNode into a piece of text on the page. Which you do not want.