External UML diagrams were no longer rendered
A recent bugfix removed the automatic passing of SpanNodes to the
directives in case of a single line directive. But this breaks the Uml
Directive because it did not look at the $data variable but the node's
value. This change will have it look at the data again and load the file