Images were not embedded when generating guides
Images are copied over when the Asset filter is used in a twig template.
Although you could debate whether it should not be a reference and
copied over in a single pass, the issue arose that the destination file
system with the wrong path was provided as source for the image
location.

This is incorrect, but it also showed that there is room for improvement
in how we deal with paths in guides. This is not addressed in this PR,
but merely that the origin filesystem is passed