False and Null are missing their first letter in a constant's value
We attempted to compensate for an earlier report (issue 2421) by undoing
the adding of slashes by PHP-Parser's PrettyPrinter. However, this had
the unintended side-effect of stripping the first slash when a boolean
or null was provided and subsequently that was interpreted as a unicode
character.

I have removed the un-escaping of the slashes by PHP-Parsers
prettyprinter, reintroducing double slashes in some instances, as this
will ensure that values are presented correctly.

While it may seem odd that a single slash in a single quoted string is
now converted into a double-slash; this notation is technically more
correct as the php documentation states that even in a single quotes
string, a double slash is intended to represent a literal backslash.

See
https://www.php.net/manual/en/language.types.string.php#language.types.string.syntax.single

Although, in code, a single backslash is also used literally; this means
the output by phpDocumentor is not incorrect; albeit verbose perhaps.