Extract Table parsing into a Subparser
In the previous commits, I have been extracting subparsers from the
DocumentParser in Guides. This time, it was time for the Table parsing.

This parser has one oddity compared to the other ones, its parse method
does not return when it resets the state to BEGIN. I am not sure yet
whether this is a bug that is becoming visible now, or some weird
workaround. Just to be sure, I left the weird bit in.