Extract Title handling to a Subparser
Titles in RestructuredText are aware of where and how they are used in a
document. This means that a lot of information is passed back and forth
and it makes for a special case in the flush. In future commits, we can
try to reduce the 'customness' even more and also see how we can reduce
the amount of state. But for now, I tried to keep it simple and just
move things around.

The openTitleNodes array has been changed into an ArrayObject so that it
could be passed by reference and so that the build function could
manipulate it. This is considered bad design, as it opens up far too
much parent state, but to prevent this change from becoming too big I
have kept it like this.