Extract initialisation of Directives into a Subparser
As a first step into migrating Directive handling to a subparser, I have
created the Subparser but limited its interactions to the creation/init
part.

Since the implementation currently creates a Directive Handler and
assigns that to a property in the DocumentParser, I can get away with
this slim step. Subsequent commits will migrate more into the Subparser.