Subparsers become Builder objects, Directive can now accept ContentBlocks
In this change, I had to make a couple more extensive changes in order
for Directives to work. Directives are treated as a sort of pushdown
system where as a property directiveParser has a non-null value; then
the next Code node is added as a ContentBlock.

This too should be changed in a future commit to properly support nested
/ tree parsing but we need to take small steps.