Initiate change to a Recursive Descend Parser
After having reworked a series of Node types into substates, I have been
able to get a glimpse of the larger system. It seems, the original
author attempted a line-based parser that could backtrack (hence, why it
could re-parse every line until a parse was successful).

During research on how others have approached and fixed this issue, I
found references to LL(1) or Recursive Descend Parsers without
backtracking and with 1 token look ahead. When I compare that method, it
seems very suitable in this situation and requires a bit of rework to
introduce "productions". This term is taken from LL(1) parser theory to
indicate a means to start at a given token (line in our case) and
construct a product or AST node by moving the cursor forward.

This deviates from the original implementation where the movement of the
cursor was driven by a foreach statement. By changing this to a while
statement it is not a problem is productions themselves move the cursor
forward.