Extract the Title interpretation into a Rule
The title is tightly coupled to sections, which are structural
elements, and through this it means that we have to rework the parser a
bit. In this change, I have promoted the DocumentParser to a Production
Rule as a first step towards nested production rules and I have reworked
parts so that section handling is actually done through the TitleRule.

This is not the final phase, but I had to reduce the complexity of the
refactoring.

After this commit, the parser is in a broken state unfortunately because
it is no longer possible to use a strangler pattern to keep it working
with old and new components. Subsequent commits will restore
functionality by moving more and more pieces of the parser into
production rules.