Extract Paragraph handling into a Production Rule
During implementation, I found out I had to tweak parts of the
recognition of characters for quote rules because of issues with
whitelines. Since a QuoteBlock recreates a parser and passes its
contents through it (to be changed), it ended up in an infinite loop
when the content merely contained an empty string.