Migrate Directive handling to a Production Rule
As the next step in finalizing the design of the parser for Restructured
Text, I am migrating directive handling to a production rule. This is
also a good example how we can re-use other Production Rules to get a
compound rule.

The implementation is a bit rough, and there are one or two bugs
remaining (I lost the last line of text in a note?) but it is a good
starting point