Lists are not rendered correctly
Due to a mistake in the ListRule, it did not collect the right amount of
lines. This is now corrected, even though rendering is still off because
ListLines are not interpreted like block quotes are.

By committing this now, it is at least an improvement and once I have
time again I can look at the changes made by Wouter in another fork of
the Doctrine Parser