Highlight active page in menu
The destinationPath twig variable did not contain the same content for
both the API docs and the Guide rendered docs. In this commit, I have
mostly fixed that (the API contains a leading slash, I need to fix that
later) and by doing so, I can compare the URL of the current page to
menu items.

This comparison will allow me to set a CSS class '-active' on menu's
making it show bold when you are on that specific page.

This feature still has one limitation though, it will not highlight
menu items that you are viewing indirectly. Think of classes, when
you open the class documentation there will be no highlight because
only namespaces are shown in the menu.

This limitation would be nice to fix in a future change, for now it
would mean too much of an impact on how we generate the menu.