Introduce Twig element documentation overview
In our templates we use custom twig variables, functions and filters,
these need to be described in detail. But as a first step I have created
a listing page that resembles that of the Twig documentation.

This experiment allowed me to introduce the container directive and
update a few bits and bobs on list styling. Previously there has been a
'div' directive that does the same thing; but the container directive is
the canonical directive described in the docutils documentation.