Restore 'raw' directive
The raw directive is implemented but, for an unknown reason, it shares
the same template as the CodeNode. In that template, raw was disabled
and this caused the raw directive to stop functioning.

In issue #3024 it was requested to restore this feature and that is what
I did in this change. There are however a couple of extra comments that
this should be refactored into its own structure; but a bit more effort
would be needed because there is probably a reason why the CodeNode and
RawNode share the same rendering process.

Fixes #3024