Extract reference handling to a new service
In order to reduce the need for the environment service and decouple all
parts in guides from it; I have moved everything related to references
to a separate service.

The intention was to fully decouple from the Environment service; but
another one or two passes are needed to accomplish that. The Environment
service contains relations and methods used internally in references
that need to be refactored away in subsequent commits.