Moving uses of Environment out of constructors
The Environment class is stateful and changes for each rendered file.
Many services rely on this to be passed through on the constructor, but
this prevents is from making these flyweight services and from re-using
those elsewhere in the code.

With the changes in this commit, I am reorganizing how the Environment
is passed along to be less focused around the constructor but instead
the moment when it is used.