Reference tracking should be a regular service
During the remodelling of the ReferenceRegistry I found that it is
actually tracking state for another system, the dependencies.

Dependencies are, as I see it now, part of a document and by moving
those into the DocumentNode the ReferenceRegistry, now ReferenceBuilder,
can act as a truly flyweight service that, based on the Document as
scope, it can now track which dependencies there are and directly
register them onto the DocumentNode.