Simplify Parser(Handler)
The Guides parser instantiation is a complicated process due to the
amount of adhoc instantiation. With these changes, I am moving most of
the instantiation into the Service Container.

The expected benefits is a simpler parser, and when we extract this into
a separate library that it is more flexible to use.