Improve writer collection implementation
The writer collection was extending \ArrayObject which caused errors on
php 8.1 since the return types were not matching the php return types.

But when looking at the purpose of this class using a collection like this
doesn't make sense. The writer collection should be immutable after the application
was booted. So I rewrote the implementation to be more driven by the symfony service
container. And removing the extend of ArrayObject which is not needed.