Refactor Guides/Configuration away
In this change, I have streamlined that process of parsing and rendering
Guides by refactoring the Configuration class away. This class created
undesirable coupling and was redundant on top of the GuideSetDescriptor.

With these changes, I am also working towards having the output formats
as a registry / central repository so that they can be injected and/or
more easily added upon; and I am configuring the Message and Handlers to
work as an anti-corruption layer

Signed-off-by: Mike van Riel <me@mikevanriel.com>