Fixes #3047: Image and Figure options do not work
When you pass a width, height or other option to an image it did not
work. This was caused because the wrong method was called; I also
checked the figure directive and it seemed that was not handled
correctly either.

As such I have changed both to now correctly use their respective
options as described in
https://docutils.sourceforge.io/docs/ref/rst/directives.html#figure