Box doesn't include incubator in PHAR
When building a PHAR file, the incubator packages do not end up in the
PHAR. I think this is because they are path-based packages that are
usually installed as a symlink, and box doesn't include symlinks.

By passing the environment variable COMPOSER_MIRROR_PATH_REPOS, composer
will mirror the package instead of using a symlink.