Migrate CDNed resources to a compiled bundle
The default template uses a few Javascript libraries and polyfills that
were loaded from a CDN. To make phpDocumentor more stand-alone, I have
moved all these assets into a webpack build.

This webpack build only needs to be ran by contributors, as the
generated output is included in the template. This will ensure that
users of phpDocumentor do not need webpack installed.