POC resolving hooks in workpress code
This is a POC using the new metadata concept in phpDocumentor to add
extra information in your docs. It is far from usefull as many changes are
needed in the core code to support this case.
This code can be used as reference code to see what we need to change in phpDocumentor
and it reveals missing extension points in our application.