Unit tests failed because prophecy could not be found
In the previous revert, I have undone the offending changes that caused
the unit tests to fail; in addition I have re-added the GPG dependency
in the Dockerfile to be able to run `make setup` again because that was
broken.

Finally, I did a quick cleanup in the gitignore to disallow any files
in /bin that are not specifically permitted. Too many tools dump files
in this bin folder :)