Workaround phive rate limits
When using phive, we run into rate limit issues because it does quite a
few requests. To work around this, we want to limit the number of
actions phive does by passing arguments to phive.

Unfortunately, phive forgets about most of its config, including the
location attribute, when you pass a limited set of packages to install.

This poses issues with the phpunit extensions as they _must_ be in the
correct folder; and thus we need to add a workaround to copy the phpunit
extensions into the correct folder.