Support Location attribute on templates
In this commit, I have added handling, tests and documentation for the
'location' attribute in the template configuration entry. This should
now work and attempt to load the template from the given directory on
the host's filesystem.

Because I kept running into issues with how phpstan and psalm interpret
the configuration array shape; I have also cleaned that part up and used
psalm-type definitions to provide a reusable shape. I chose not to use
phpstan-type annotations because phpstan understands the psalm
annotations but not vice versa.

Unfortunately, I discovered that we are using an antiquated version of
phpstan that does not like array-shape type aliases. This is fixed in a
newer version but I need @jaapio his input whether we version locked for
a reason.

Signed-off-by: Mike van Riel <me@mikevanriel.com>