Tightening the configuration interfaces
The work done in this commit is partly to satisfy Psalm 1.8. Not that it
is wrong, but there were quite a few improvements to be had as the
configuration section of phpDocumentor is quite dynamic. This could use
some tightening in the future, but for now these changes will help a
great deal already.