Support null versions in Since reflection
The Since Reflection object may return a null value for the version
field. The Assembler however passed this without checking onto the
Descriptor, and the descriptor requires a string value, not null.

This change will ensure the app doesn't crash and I got to add a new
test that wasn't there before!

Fixes #3227