Levenshtein argument too long
Apparently, Symfony Console uses the levenshtein function to determine
whether it can find a similarly named command in case you make a typo.
However, in some weird cases, the FirstArgument can be more than 255
characters and that same function will complain because we do some
console command switcharoo magic. Because a command should not be more
than 100 characters in length, I prevent the warnings by checking
whether FirstArgument is more than 100 characters, if so: we discard
that.