Move version number determination into Application
As it stood, there were now 2 places where the version was determined.
The Console Application class and the main Application class. Since I
need the version number to be readily available to check against, I have
moved the version number determination to the main Application class and
re-used the same methodology of the Console application class.