Code cleanup after findings
While working on invalidating the caches upon a version number change, I
found that upon changing the API Specification no cache purge happens
anymore. This mechanism should be there because a change in the API
Specification requires a full reparse of the Descriptors because the
cached state may not match the specification. This can cause weird
issues with people's output as the previously cached bits do not match
their configuration