Relative cache paths on cmdline are wrong
When a cache path is provided through the command line, it should be
relative to the current working directory. The default behaviour of
phpDocumentor is that configuration options, such as the cache path, are
relative to the config file unless declared absolute.

As such, when interpreting the cache-path command line option, we
convert it to an absolute path if it is relative. This will ensure that
the configuration interpreter will leave it alone.

Fixes #3158