When typing an array, show that in docs
Properties that are arrays need to be typed using a DocBlock; but when a
typehint _and_ a DocBlock was specified; the contents of the DocBlock's
var tag was gone.

This change will introduce similar behaviour to the ArgumentAssembler by
overriding the type of the Descriptor from that of the found var tag.

I did bake in that it will only do that when there is 1 var tag; when
there are none or multiple it will not do this to prevent issues.