Line numbers show 0 with InvalidTag issues
The InvalidTag does not know its line number or location and as such the
errors overview in the generated documentation shows a line number of 0.
This is not quite helpful.

To improve the situation, I have introduced that any line number of 0
will be overridden with the containing element's line number when you
retrieve the error listing.

Fixes #3057