Windows tests keep failing
I have removed phive from the build pipeline because it keeps giving
timeouts when running the windows tests. This is a combination of phive
having too strict a timeout apparently and the Windows runners of Github
Action being slow.

Even though it would be nice to keep Phive in the pipeline, I just
cannot keep the windows tests failing much longer and the only surefire
fix is to replace phive with oldschool tooling