Windows tests keep failing
I have removed phive from the build pipeline because it keeps giving
timeouts when running the windows tests. This is a combination of phive
having too strict a timeout apparently and the Windows runners of Github
Action being slow.

Even though it would be nice to keep Phive in the pipeline, I just
cannot keep the windows tests failing much longer and the only surefire
fix is to replace phive with oldschool tooling

At the end of this change, the windows build still fails. But now it is due to 
a legitimate failure. As such, I have kept the setting on that will allow failures 
on windows; this way I can fix those failures in a subsequent commit