Do not show container section of TOC in classes
Classes have a method getInterfaces() with which you describe which
interfaces a Class implements, but this conflicts in the
table-of-contents template as this assumes that if there is a
getInterfaces() method it will show a section intended for namespaces,
packages and files.

This change introduces twig test functions to test for various types of
descriptors, and an extra check in the table of contents template
whether the given node is an element container.