Introduce the first value objects
During a recent call, Jaap and I talked about using imposters to be able
to have objects that represent references. This can replace the use of
multiples types in the linker and descriptors by having a Reference
object instead of a FQSEN directly injected.

The value objects allow us to be more flexible in our design, and to
return an 'undefined' state when using these imposter objects.