Restyle elements and add support for colors
To make adoption easier, I have added support for setting a color using
the command line using the settings functionality. The setting in
question is 'template.color'.

Because settings were a flat structure before and I wanted to add
multiple settings under the template umbrella I changed it to explode on
a '.' in the key.

In addition, the elements have been changed visually to be more legible
and use a different font.