Show whether a method is tagged as 'api'
In the default template, there is no indication whether an element is
tagged with the `api` tag, this change fixes that for methods at least.

Another change included in this commit is that we now properly do not
render response descriptions, including the em-dash, when there is none
to render, and when it is of type mixed without a description we also
omit rendering of the result because it doesn't add anything.