Split LinkRenderer into 3 services
The LinkRenderer is rather complicated, and when you look closer at it,
this has 2 reasons:

1. it violates the SRP and actually does 3 things
2. it violates the SRP even more because the linkrendering itself is
actually a couple of adapters in a chain of responsibility all coded
into a single method.

In this commit, I address the first issue by splitting the link renderer
into 3 services: LinkRenderer, PathBuilder and the
RelativePathToRootConverter.

These names may not be final yet as some more work needs to be done, but
it sets the stage for transforming the LinkRenderer into a chain of
responsibility with adapters per type of object that needs to be
transformed and to extract the presentation style to a decorator or
another type of adapter.