Allow linking to guides using doc:// scheme
This is what I was working towards, making links from the
api documentation to the co-generated guides. The code needs to be
cleaned up and needs quite a few tests, but we are getting there.

Also, the descriptions from see tags are not used as labels. So there's
that to fix