Split off AbstractListAdapter
To me, this was the missing key for why there was still something messy
going on in the LinkRenderer; the AbstractList exception in the
LinkAdapter was really just another adapter.

With this, I can start writing tests; and even though there is still
room for improvement. It will have to do for now