Warn when multiple versions and/or sets of each type is given
As part of the effort to support multiple versions and sets of
documentation we first need to make sure there can only be one version
and one set of documentation per type.

This sounds counter-intuitive, but adding this restriction will help the
user experience, and it will allow us to make assumptions on the number
of versions and sets while proxying data from one side of the
application to another. Specifically the 'indexes' in
ProjectDescriptor can be proxied to the first version's first Api Set