Move Files listing from Project to Documentation Sets
As part of the effort to introduce support for multiple versions and
components, we need to move data around to go from the project (all
versions and components) to Documentation Sets (which is bound to a
single version and component).

To be fair, this change does not move everything around; just where
files are stored at the moment. This is done to keep the changesets
small. In subsequent work, we can change the usage of the get and
setFiles to start pointing to the DocumentationSet.