Move indexes from Project to Documentation Sets
In this change, I have moved the indexes that are build during the
parsing of a documentation set from the project descriptor to the
Documentation Set descriptor. This is another step in making multiple
components and versions possible as the data is now with a documentation
set instead of globally.

The findElement method has not been moved to the generic
DocumentationSet because it relies on an FQSEN, which is an API specific
construct; and as such it is now moved to the ApiSetDescriptor
specifically