Extract API Documentation Set into a sub-pipeline
With this change, I have extracted the parsing of each API set into a
sub-pipeline. This means that we can now parse multiple sets of API
Documentation instead of just the first.

However, it doesn't work quite yet. As the files are still added onto
the ProjectDescriptor. This change merely changes the process, but
shouldn't affect the application as long as the guard that only 1 API
Set may be present is in place.

In a subsequent change, we will be modifying the parsing logic itself to
work more with the Documentation Set descriptors instead of the
ProjectDescriptor