Post self-review clean up
While reviewing, I noticed that a) I had added a method to the
ApiSetDescriptor that I really do not need anymore, and b) the Parser
Payload is no longer really used.

I have now replaced the Parser Payload with its parent (the normal
Payload) and found that the stage to transform the regular Payload to
the Parser payload was unnecessary; so I cut it out to simplify the
process.