Use ApiSetDescriptors instead of Specification
During the parsing phase, I had used ApiSpecifications instead of the
ApiSetDescriptor because I believed the ApiSetDescriptors were created
later. I found that the ApiSetDescriptors are built in the
InitializeBuilderFromConfig stage, meaning I can use them in the parse
stages.

This will simplify the change where files and indexes are directly set
in the ApiSetDescriptor instead of on the ProjectDescriptor because the
ParseFiles stage now has direct access to the ApiSetDescriptor