Populate Api Set directly instead of the Project
In this change, I have changed the way the ParseFiles stage works by
having it directly push the parsed files into the ApiSetDescriptor
instead of the ProjectDescriptor.

Since the ProjectDescriptor's Files and Indexes collections are proxies
for the first ApiSet at the moment, nothing changes in the way
phpDocumentor works. This change is again moving forward with the plan
to remove the getFiles and getIndexes from the ProjectDescriptor. As
soon as that is gone, we have successfully decoupled documentation sets
from the main project and be able to support multiple of them.