Move compiling into a pipeline
The compiler was introduced before we had pipelines, and to make
maintenance simpler and more uniform I had moved the compiling steps
into a pipeline.

In addition, an extra decorator was added that will log the start, end
and execution time for each stage. This will be visible when running
phpDocumentor in verbose mode.