Change writers to be performed on a per-documentation-set basis
We are inching closer to supporting the whole documentation
pipeline per documentation set instead of a per-project level.

This is one of the final steps to make the architecture work in this
way.

Once this is solved, we can start experimenting with enabling
multi-documentation-set rendering and discovering the issues stemming
from that.