Initial setup for views
The templates currently rely on the project descriptor, but we want to
get rid of the getFiles and getIndexes methods in the Project
Descriptor.

Instead of changing all the templates, I am looking at introducing a
View layer in phpDocumentor so that we can replace the direct use of
these top level descriptors with a set of views.

The idea behind this is that in the long run, views are aware of the
type of documentation set and can change their output depending on this.

But, give it a couple more commits to crystalize. I am also not too sold
on it being a static factory method at the moment..