Enable Code Coverage when locally running PHPUnit
I occasionally want to add a unit test from the backlog of unwritten
tests. To do this, I need the code coverage report but that was disabled
a while ago.

To reintroduce the code coverage in a non-intrusive way I have
introduced a new build stage in the Dockerfile that will install pcov,
and I make sure the production builds stop at the prod target.

The added benefit from this is that we use the phpdoc container to run
the tests in, meaning that if a test fails in the container it will
surely fail in production as both environments are identical