Fix: resolves issue with varidict arguments
This fixes an issue with varidict arguments, if a docblock contained
an param with a varidict argument this was overwritten, however
this should only be done when the argument itself is not available.

This feature enables people to create a virtual param this is not
defined in the function header, for example when they use function_get_args.