Feat: Add filter for var tags on groups
We do allow multiple var tags in a docblock, this is usefull
when multiple properties or constants are defined in the same line.
This patch will introduce filtering by variable name on those tags.
this will allow users to define multiple var tags and reference specific
properties or constants to document them.