Add support for #[Deprecated]
The new attribut was added in php 8.4 and we no now read it
as a normal docblock item. This way our users can use the new
attribute to deprecate their methods, classes and properties
and use those in the documentation as well.

To make deprecations better visual they are new having their own
component which will make it possible to customize the way deprecations
are shown.