[FEATURE] filter empty namespace
Namespaces that are completely empty after filtering elements are
now hidden from the documentation. This mostly happens when a namespace
only contains internal elements. As we filter those by default
empty namespaces are the result. This makes it more confusing to
browse the docs.

If a tree of namespaces is empty all parent namespaces are also include.