Resolve issue with guide rendering in subpath
When rendering guides we do this to a subpath, however
the resolving of images happens based in on the target location
of files. This results in our context to issues.

I introduced a new UrlGenerator for guides that takes care of this
in the context of phpDocumentor. This also allowed me to remove
some lines that were added in the beginning of guides to fix relative
path rendering.

fixes #3746