Add extension finder
The extension finder allows users to install extensions using composer.
As an alternative to regular packages we do support phar extensions,
which can be scoped to avoid conflicts with the main phpdocumentor application.

The finder script is trying to avoid collisions between phpdocumentor
and the classes within the project being processed. This might be the
case when packages used by phpDocumentor are also used in the project.